core_drf/pyproject.toml
[project]
name = "core-drf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.16.1",
]
